Moved sample_from_distributions from helpers.py to numerical/helpers.py .
